+++ src/App.css
@@ -1,13 +1,11 @@
-/* General styling */
 body {
   margin: 0;
   font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
-  background-color: #f4f7fa; /* Light background for better contrast */
-  overflow-x: hidden; /* Prevent horizontal overflow */
-  padding: 20px; /* Add some space around the body */
+  background-color: #f4f7fa;
+  overflow-x: hidden;
+  padding: 20px;
 }
 
-/* Heading section */
 .heading {
   display: flex;
   align-items: center;
@@ -18,18 +16,17 @@ body {
   background-color: rgb(66, 168, 134);
   padding: 15px;
   color: white;
-  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1); /* Soft shadow for depth */
-  margin-bottom: 30px; /* Space below the heading */
+  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
+  margin-bottom: 30px;
   border-radius: 8px;
 }
 
 .logo {
   height: 40px;
   margin-right: 10px;
-  flex-shrink: 0; /* Prevent shrinking */
+  flex-shrink: 0;
 }
 
-/* Main container */
 .container {
   display: flex;
   justify-content: center;
@@ -41,7 +38,6 @@ body {
   margin-top: 20px;
 }
 
-/* Search and theme toggle section */
 .search-theme-container {
   display: flex;
   justify-content: center;
@@ -115,7 +111,6 @@ body {
   background-color: #e0e0e0;
 }
 
-/* Table Styling */
 table {
   width: 100%;
   max-width: 800px;
@@ -153,13 +148,11 @@ a:hover {
   text-decoration: underline;
 }
 
-/* Avatar Styling */
 .avatar {
   border-radius: 50%;
   margin-right: 10px;
 }
 
-/* Profile Section */
 .profile-info {
   display: flex;
   justify-content: center;
@@ -178,7 +171,6 @@ a:hover {
   border-radius: 50%;
 }
 
-/* Tab Navigation */
 .TabNavigation {
   display: flex;
   justify-content: center;
@@ -210,7 +202,6 @@ a:hover {
   text-decoration: none;
 }
 
-/* Responsive Design */
 @media (max-width: 768px) {
   .container {
     flex-direction: column;
